Fixes for login form style
The built in form does not match the bootstrap css selectors. There were
two options. 1. Override the built in form generator to match bootstrap.
2. Hack (copy paste from bootstrap) a couple of css rules to match what
we have. 1. seems cleaner but has larger code impact. 2. is easy and
only css is changed. The login form has id="login" to match on. I went
for no 2.