Per image gallery control to image accessibility
Image accessibility is normally set on the Image edit form. However
this is rather inconvenient after acquiring a bunch of orphans if that
had been just uploaded with the default 'Public' accessibility. Now we
can set accessibility from the gallery images page that has a per image
checkbox or an option to select all.