Fix Exif exception when no Exif info is found
The Exif.fromFile threw an exception which caused the image to be
uploaded but not put into the database. This was reported back to the
used as an failure in the upload but as the image files were present it
also caused an md5 clash.
The fix catches the Exif exception and returns Nothing if it was raised.